+++ packages/client-multiplexer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @agentick/client-multiplexer
 
+## 0.14.64
+
+### Patch Changes
+
+- @agentick/client@0.14.64
+
+## 0.14.63
+
+### Patch Changes
+
+- @agentick/client@0.14.63
+
 ## 0.14.62
 
 ### Patch Changes