+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @agentick/nestjs
 
+## 0.14.64
+
+### Patch Changes
+
+- @agentick/gateway@0.14.64
+
+## 0.14.63
+
+### Patch Changes
+
+- @agentick/gateway@0.14.63
+
 ## 0.14.62
 
 ### Patch Changes