+++ packages/sandbox-secure-exec/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @agentick/sandbox-secure-exec
 
+## 0.14.64
+
+### Patch Changes
+
+- Updated dependencies [71ce3e0]
+  - @agentick/core@0.14.64
+  - @agentick/sandbox@0.14.64
+
+## 0.14.63
+
+### Patch Changes
+
+- @agentick/core@0.14.63
+- @agentick/sandbox@0.14.63
+
 ## 0.14.62
 
 ### Patch Changes